+++ packages/article-comments/article-comments.stories.native.js
@@ -1,0 +1,4 @@
+import { showcaseConverter } from "@times-components/storybook";
+import showcase from "./article-comments.showcase.native";
+
+showcaseConverter(module, showcase);