+++ packages/provider-test-tools/src/index.js
@@ -1,5 +1,6 @@
 import schema from "@times-components/schema/schema.json";
 import article from "./article";
+import articleComments from "./article-comments";
 import articleListNoImages from "./fixtures/article-list-no-images.json";
 import articleListWithImages from "./fixtures/article-list-with-images.json";
 import author from "./fixtures/author.json";
@@ -31,6 +32,7 @@ const makeMocks = mm(schema);
 export * from "./helpers";
 export {
   article,
+  articleComments,
   authorProfile,
   clientTester,
   fixtures,