+++ packages/article/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="6.0.3"></a>
+## [6.0.3](https://github.com/newsuk/times-components/compare/@times-components/article@6.0.2...@times-components/article@6.0.3) (2018-11-07)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
 <a name="6.0.2"></a>
 ## [6.0.2](https://github.com/newsuk/times-components/compare/@times-components/article@6.0.1...@times-components/article@6.0.2) (2018-11-06)